extract register controller
