+++ src/Http/Controllers/User/UserController.php
@@ -2,20 +2,13 @@
 
 namespace Statamic\Http\Controllers\User;
 
-use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\MessageBag;
-use Illuminate\Validation\ValidationException;
 use Statamic\Auth\ThrottlesLogins;
-use Statamic\Events\UserRegistered;
-use Statamic\Events\UserRegistering;
-use Statamic\Exceptions\SilentFormFailureException;
 use Statamic\Facades\User;
 use Statamic\Http\Controllers\Controller;
 use Statamic\Http\Requests\UserLoginRequest;
 use Statamic\Http\Requests\UserPasswordRequest;
 use Statamic\Http\Requests\UserProfileRequest;
-use Statamic\Http\Requests\UserRegisterRequest;
 
 class UserController extends Controller
 {
@@ -45,42 +38,6 @@ class UserController extends Controller
         Auth::logout();
 
         return redirect(request()->get('redirect', '/'));
-    }
-
-    public function register(UserRegisterRequest $request)
-    {
-        $user = User::make()
-            ->email($request->email)
-            ->password($request->password)
-            ->data($request->processedValues());
-
-        if ($roles = config('statamic.users.new_user_roles')) {
-            $user->explicitRoles($roles);
-        }
-
-        if ($groups = config('statamic.users.new_user_groups')) {
-            $user->groups($groups);
-        }
-
-        try {
-            if ($honeypot = config('statamic.users.registration_form_honeypot_field')) {
-                throw_if(Arr::get($request->input(), $honeypot), new SilentFormFailureException);
-            }
-
-            throw_if(UserRegistering::dispatch($user) === false, new SilentFormFailureException);
-        } catch (ValidationException $e) {
-            return $this->userRegistrationFailure($e);
-        } catch (SilentFormFailureException $e) {
-            return $this->userRegistrationSuccess(true);
-        }
-
-        $user->save();
-
-        UserRegistered::dispatch($user);
-
-        Auth::login($user);
-
-        return $this->userRegistrationSuccess();
     }
 
     public function profile(UserProfileRequest $request)
@@ -114,46 +71,6 @@ class UserController extends Controller
     public function username()
     {
         return 'email';
-    }
-
-    private function userRegistrationFailure($validator)
-    {
-        $errors = $validator->errors();
-
-        if (request()->ajax()) {
-            return response([
-                'errors' => (new MessageBag($errors))->all(),
-                'error' => collect($errors)->map(function ($errors, $field) {
-                    return $errors[0];
-                })->all(),
-            ], 400);
-        }
-
-        if (request()->wantsJson()) {
-            return (new ValidationException($validator))->errorBag(new MessageBag($errors));
-        }
-
-        $errorResponse = request()->has('_error_redirect') ? redirect(request()->input('_error_redirect')) : back();
-
-        return $errorResponse->withInput()->withErrors($errors, 'user.register');
-    }
-
-    private function userRegistrationSuccess(bool $silentFailure = false)
-    {
-        $response = request()->has('_redirect') ? redirect(request()->get('_redirect')) : back();
-
-        if (request()->ajax() || request()->wantsJson()) {
-            return response([
-                'success' => true,
-                'user_created' => ! $silentFailure,
-                'redirect' => $response->getTargetUrl(),
-            ]);
-        }
-
-        session()->flash('user.register.success', __('Registration successful.'));
-        session()->flash('user.register.user_created', ! $silentFailure);
-
-        return $response;
     }
 
     private function userProfileSuccess(bool $silentFailure = false)
